requirements.txt
# Django bootstrap v5
django-bootstrap-v5

# Django Widget Tweaks
django-widget-tweaks